+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 Node90 Firmware
 
 # 1.6.0
-- add pwm, i2c, io_expander
+- add pwm, i2c, io_expander, dac
+- add actuator mode, dim_els, cuv_enable
 
 # 1.5.0
 - add actuators, sensors, policy, and event resources